pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant_rec"
version = "0.1.0"
description = "AI-powered Zomato-style restaurant recommender"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1",
    "pyarrow>=15.0",
    "datasets>=2.18",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "groq>=0.11.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]

[project.scripts]
ingest-zomato = "restaurant_rec.phase1.ingest:main"

[tool.setuptools.packages.find]
where = ["src"]
